Create Attempt 1
Line 39 is where I am having issues. I pass the board into the method called "gameStatusDraw" -- to determine whether the game is a draw. BUT it seems that the result always comes back that its a draw, regardless.  

I am having trouble figuring out the return statements within the method on line 109. The Method wants me to have a return statement for it. But I do have it in line 118 and 127.

But it wants it in 130 before it leaves....